d/rules: set GO111MODULE to auto
In Go 1.16, which is the default in Hirsute now, the packages are built
in module-aware mode. Since containerd does not have a go.mod file in
its source tree it FTBFS. Setting GO111MODULE to auto we can have the
previous behavior which is enable module-aware mode only if the go.mod
file exists.